Fixed up clean conditions

+++ test/install.sh
@@ -14,7 +14,7 @@ fixture_git_init() {
 }
 
 # A normal victorious-git installation
-rm -r ~/.config/victorious-git/
+if test -d "~/.config/victorious-git"; then rm -r ~/.config/victorious-git/; fi
 "$repo_dir/bin/install.sh"
 
   # when initializing a git directory
@@ -41,7 +41,7 @@ rm -r ~/.config/victorious-git/
     fi
 
 # A symlink victorious-git installation
-rm -r ~/.config/victorious-git/
+if test -d "~/.config/victorious-git"; then rm -r ~/.config/victorious-git/; fi
 "$repo_dir/bin/install-symlink.sh"
 
   # when initializing a git directory
